examples/hellobp/requirements.txt
bpolicy==0.3.0
python-memcached==1.53
web.py==0.37
